Education component
Still need to make many improvements. For desktop version I need to adjust the amount of pixels the arrows scroll horizontally within the Course component to a dynamic value according to the window size. Need to set the progress bar to stay in a fixed position. Visually, I need to improve spacing, font-size, arrows and progress bar (virtually everything...)