Giphy APi
Added Giphy API. Creation of env file to store Api key. Added .env to gitignore to avoid disclosing it (still need to figure out how to protect this key during deployment given the fact that this gets injected into the code when it gets bundled up and shipped to the browser, so it can be accessed simply by inspecting the website).  Still need to improve overall styles.